18GB: Fixed token-laying errors (These were coming from some botched inheritance order on the track_and_token step - now tidied up)
